# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Vida Refrigeração
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, inactive]
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [pessoa_fisica, pessoa_juridica]
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/search:
    get:
      operationId: searchClients
      tags: [clients]
      summary: Search active clients (autocomplete)
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        clientType:
          type: string
          enum: [pessoa_fisica, pessoa_juridica]
        document:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        cep:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        addressNumber:
          type: string
          nullable: true
        addressComplement:
          type: string
          nullable: true
        neighborhood:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        origin:
          type: string
          enum: [manual, conta_azul]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - clientType
        - document
        - status
        - origin
        - createdAt
        - updatedAt
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        clientType:
          type: string
          enum: [pessoa_fisica, pessoa_juridica]
        document:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        cep:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        addressNumber:
          type: string
          nullable: true
        addressComplement:
          type: string
          nullable: true
        neighborhood:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        origin:
          type: string
          enum: [manual, conta_azul]
      required:
        - name
        - clientType
        - document
    UpdateClientBody:
      type: object
      properties:
        name:
          type: string
        clientType:
          type: string
          enum: [pessoa_fisica, pessoa_juridica]
        document:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        cep:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        addressNumber:
          type: string
          nullable: true
        addressComplement:
          type: string
          nullable: true
        neighborhood:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        origin:
          type: string
          enum: [manual, conta_azul]
...